+++ noodles-csi/CHANGELOG.md
@@ -6,7 +6,7 @@
 
   * csi/async/io: Add async indexed reader (`csi::r#async::io::IndexedReader`).
 
-  * csi/async/io/reader: Add async query reader (`csi::r#async::io::Query`).
+  * csi/async/io: Add async query reader (`csi::r#async::io::Query`).
 
 ## 0.51.0 - 2025-08-25