+++ noodles-csi/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.53.0 - 2025-12-11
 
 ### Added